added the throw vs place mechanism
I will explain the controls in person but for now, it is in. What is an issue to me is that the custom friction hasn't been implemented yet and I have to do that soon so the character won't fall down corners. this will also create the ledge forgiveness we talked about.

What I also added was a UI element that works with the current talisman mode
- Bind
- Emotion

and when in emotion, it switches between Joy and Lonely

an altar, which is just an object with the altar class on it, can switch the emotion of the room in the Emotion type of talisman. this rn is just a quick fix and can be changed.

Since I was too lazy to copy the material being transparent, the cubes have a UI element now that says which emotion they are currently in