added coyote time and comments
I added comments on the script for the job. specifically what each value is for. I also changed the check type from a raycast to a spherecast so that if the player is stuck in a slit, they should still be able to jump. if that won't help as well, We can implement something like a "0 velocity jump" when the player is stuck somehow but doesn't fall down.