fixed slope velocity problem
the player now doesn't fall down slowly when the slope angle is beneath the accepted value. please make sure you don't have an angle that is above it, as it will not accept that as ground etc.  additionally, the player can't cheat their way up a slope now with enough velocity, as there is a stronger force pushing them down a slope when they shouldn't be able to walk on that surface